refactor: extract issue template create logic
