+++ frontend/src/plugins/issue/logic/initialize/standard.ts
@@ -1,0 +1,45 @@
+import { TemplateType } from "@/plugins";
+import { IssueCreate, MigrationType } from "@/types";
+import { findDatabaseListByQuery } from "../common";
+import { ESTABLISH_BASELINE_SQL, VALIDATE_ONLY_SQL } from "./constants";
+import { IssueCreateHelper } from "./helper";
+import { BuildNewIssueContext } from "../types";
+
+export const buildNewStandardIssue = async (
+  context: BuildNewIssueContext
+): Promise<IssueCreate> => {
+  const helper = new IssueCreateHelper(context);
+
+  await helper.prepare();
+
+  // standard single-stage or multi-stage issue is generated by specifying
+  // databaseId for each stage
+  const databaseList = findDatabaseListByQuery(context);
+  const templateType = context.route.query.template as TemplateType;
+  let migrationType: MigrationType = "MIGRATE";
+  if (templateType === "bb.issue.database.data.update") {
+    migrationType = "DATA";
+  }
+  if (templateType === "bb.issue.database.schema.baseline") {
+    migrationType = "BASELINE";
+  }
+  const statement =
+    migrationType === "BASELINE" ? ESTABLISH_BASELINE_SQL : VALIDATE_ONLY_SQL;
+  helper.issueCreate!.createContext = {
+    migrationType,
+    updateSchemaDetailList: databaseList.map((db) => {
+      return {
+        databaseId: db.id,
+        databaseName: db.name,
+        statement,
+      };
+    }),
+  };
+
+  await helper.validate();
+
+  // clean up createContext for standard issues
+  helper.issueCreate!.createContext = {};
+
+  return helper.generate();
+};
